refactor: rename .txt to .text
