+++ dynamiq/nodes/converters/__init__.py
@@ -4,5 +4,5 @@ from .html import HTMLConverter
 from .llm_text_extractor import LLMImageConverter, LLMPDFConverter
 from .pptx import PPTXFileConverter
 from .pypdf import PyPDFConverter
-from .txt import TextFileConverter
+from .text import TextFileConverter
 from .unstructured import UnstructuredFileConverter
